clean: move refresh BrigtID connection button action to userProfile context
